Fix various Room bugs
- Prevents players from ending on Dead-end room
- Tracks whether dead-end treasure has already been looted
- Moves slow text functionality to print_formats file
- Moves all app-wide constants to helpers file
