+++ lib/helpers.py
@@ -1,0 +1,6 @@
+
+
+## CONSTANTS
+WINDOW_WIDTH = 48
+VICTORY_LEVEL = 6
+DEBUGGING = True
