+++ lib/prints/print_formats.py
@@ -1,4 +1,6 @@
-WINDOW_WIDTH = 36
+import time
+import re
+from helpers import DEBUGGING, WINDOW_WIDTH
 
 
 """
@@ -91,13 +93,29 @@ def print_header(header):
     print("|" + "{:^{}s}".format(header, WINDOW_WIDTH - 2) + "|")
     print("+" + "-" * (WINDOW_WIDTH - 2) + "+")
 
+# prints an empty line, "types" out the text, prints another empty line
+# TALK ABOUT WHERE TO PUT THIS/IF WE EVEN WANT IT.
+def slow_text(text, delay=0.03):
+    if DEBUGGING:
+        print(text)
+    else:
+        print()
+        sentences = re.split(r"(?<=[.!?])\s+", text)
+
+        for sentence in sentences:
+            for char in sentence:
+                print(char, end="", flush=True)
+                time.sleep(delay)
+            print()
+            time.sleep(0.5)
+
 
 if __name__ == "__main__":
     looping = True
     while looping:
         choice = print_menu(test_menu)[0]
         if choice == "1":
-            menu = sign_up_menu_dict
+            pass
         elif choice == "2":
             menu = main_menu_dict
         elif choice == "3":
