+++ lib/prints/print_outs.py
@@ -1,0 +1,7 @@
+## For defining variables of single-string messages
+
+
+
+## TREASURE ROOM
+find_sword = "You find a new sword! Your attack is increased by 2. After taking the sword, you return to the previous area."
+find_health_potion = "You find a healing potion! Your health is increased by 1. After taking the potion, you return to the previous area."
